Refactoring
Remove the IEnumerable check from the ResponseWrapper
Move Pagination classes to their own top-level namespace
