+++ Hackney.Shared.Finance/Boundary/Response/MetadataModel.cs
@@ -1,10 +1,11 @@
-﻿using Newtonsoft.Json;
+﻿using Hackney.Shared.Finance.Pagination;
+using Newtonsoft.Json;
 
 namespace Hackney.Shared.Finance.Boundary.Response
 {
     public class MetadataModel
     {
         [JsonProperty("pagination", NullValueHandling = NullValueHandling.Ignore)]
-        public Pagination Pagination { get; set; }
+        public PagingDetails Pagination { get; set; }
     }
 }
