+++ Hackney.Shared.Finance/Boundary/Response/ResponseWrapper.cs
@@ -1,7 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-
-namespace Hackney.Shared.Finance.Boundary.Response
+﻿namespace Hackney.Shared.Finance.Boundary.Response
 {
     public class ResponseWrapper<T> where T : class
     {
@@ -11,10 +8,6 @@ namespace Hackney.Shared.Finance.Boundary.Response
         {
             get
             {
-                if (this.Value is IEnumerable<object>)
-                {
-                    return !(this.Value as IEnumerable<object>).Any();
-                }
                 return this.Value is null;
             }
         }
