+++ Hackney.Shared.Finance/Pagination/PagingParametersDomain.cs
@@ -1,5 +1,4 @@
-﻿
-namespace Hackney.Shared.Finance.Boundary.Response
+﻿namespace Hackney.Shared.Finance.Pagination
 {
     public class PagingParametersDomain
     {
